+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Unit tests for HTML-rewriting logic.
 ### Changed
 * Do not force a newline after the doctype in `HTMLRewriter.handle_decl()`.
 * Moved `test_extracted_archive_display` from the unit tests to a separate script.